Add position demo and start work on model matrix
Added position_demo.html to show usage of setTranslate and setScale.

Tried to modify demo so that zooming would be centered on the pointer,
but this turned out to be awkward (or I'm not thinking that clearly today)
I've been thinking about supporting a general transformation matrix, so
started the implementation on that.  Added simplematrix.js with some basic
functionality, and added a model transform matrix to the vertex shader.